+++ tests/test_metadata.py
@@ -114,6 +114,7 @@ expected = {
         "gha_utils/__main__.py",
         "gha_utils/changelog.py",
         "gha_utils/cli.py",
+        "gha_utils/labels.py",
         "gha_utils/mailmap.py",
         "gha_utils/matrix.py",
         "gha_utils/metadata.py",