+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Add `gha-utils labels` command to dump bundled label configuration files.
+
 ## [5.1.0 (2026-01-23)](https://github.com/kdeldycke/workflows/compare/v5.0.1...v5.1.0)
 
 - Add new `gha-utils release-prep` command to consolidate all release preparation steps.